+++ parser_library/src/expressions/conditional_assembly/ca_operator_binary.h
@@ -45,7 +45,7 @@ public:
 
 // binary CA operators - + - * / .
 template<typename OP>
-class ca_basic_binary_operator : public ca_binary_operator
+class ca_basic_binary_operator final : public ca_binary_operator
 {
 public:
     ca_basic_binary_operator(ca_expr_ptr left_expr, ca_expr_ptr right_expr, range expr_range)
@@ -59,7 +59,7 @@ public:
 };
 
 // function binary CA operators - AND, SLL, OR, ...
-class ca_function_binary_operator : public ca_binary_operator
+class ca_function_binary_operator final : public ca_binary_operator
 {
 public:
     ca_function_binary_operator(ca_expr_ptr left_expr,
@@ -68,6 +68,8 @@ public:
         context::SET_t_enum expr_kind,
         range expr_range,
         context::SET_t_enum parent_expr_kind = context::SET_t_enum::UNDEF_TYPE);
+
+    undef_sym_set get_undefined_attributed_symbols(const evaluation_context& eval_ctx) const override;
 
     void resolve_expression_tree(ca_expression_ctx expr_ctx, diagnostic_op_consumer& diags) override;